requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-cmake-test==1.3.6
ament-copyright==0.12.9
ament-cppcheck==0.12.9
ament-cpplint==0.12.9
ament-flake8==0.12.9
ament-index-python==1.4.0
ament-lint==0.12.9
ament-lint-cmake==0.12.9
ament-package==0.14.0
ament-pep257==0.12.9
ament-uncrustify==0.12.9
ament-xmllint==0.12.9
angles==1.15.0
appdirs==1.4.4
argcomplete==1.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
bond==3.0.2
Brotli==1.0.9
builtin-interfaces==1.2.1
catkin==0.8.10
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==23.3.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.15.1
colcon-defaults==0.2.8
colcon-devtools==0.2.4
colcon-installed-package-information==0.1.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.1
configobj==5.0.6
constantly==15.1.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
docutils==0.17.1
domain-coordinator==0.10.0
dwb-msgs==1.1.13
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
flake8==4.0.1
Flask==2.0.1
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
genmsg==0.5.16
genpy==0.6.16
geometry-msgs==4.2.3
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
interactive-markers==2.3.2
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
lark==1.1.1
laser-geometry==2.4.0
launch==1.0.4
launch-ros==0.19.6
launch-testing==1.0.4
launch-testing-ros==0.19.6
launch-xml==1.0.4
launch-yaml==1.0.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.3
lxml==4.8.0
lz4==3.1.3+dfsg
map-msgs==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
message-filters==4.3.3
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
nav-2d-msgs==1.1.13
nav-msgs==4.2.3
nav2-common==1.1.12
nav2-msgs==1.1.12
nav2-simple-commander==1.0.0
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
osrf-pycommon==2.0.2
packaging==21.3
paramiko==2.9.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pexpect==4.8.0
picamera==1.13
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-qt-binding==1.1.2
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.11
rcutils==5.1.4
requests==2.25.1
resource-retriever==3.1.1
rmw-dds-common==1.6.0
robot-ros==0.0.0
roman==3.3
ros2action==0.18.7
ros2bag==0.15.8
ros2cli==0.18.7
ros2component==0.18.7
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.6
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2topic==0.18.7
rosapi==1.3.2
rosapi-msgs==1.3.2
rosbag2-interfaces==0.15.8
rosbag2-py==0.15.8
rosbridge-library==1.3.2
rosbridge-msgs==1.3.2
rosbridge-server==1.3.2
rosclean==1.15.8
rosdistro-modules==0.9.0
rosgraph==1.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
roslaunch==1.15.14
roslib==1.15.8
rosmaster==1.15.14
rosparam==1.15.14
rospkg==1.5.0
rospkg-modules==1.5.0
RPi.GPIO==0.7.1
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.2
rqt-graph==1.3.0
rqt-gui==1.1.5
rqt-gui-py==1.1.5
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.5
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.8.0
SecretStorage==3.3.1
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
service-identity==18.1.0
shape-msgs==4.2.3
simplejson==3.17.6
six==1.16.0
slam-toolbox==2.6.8
smclib==3.0.2
snowballstemmer==2.2.0
sos==4.5.6
soupsieve==2.3.1
sros2==0.10.4
ssh-import-id==5.11
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.3.2
tf2-geometry-msgs==0.25.5
tf2-kdl==0.25.5
tf2-msgs==0.25.5
tf2-py==0.25.5
tf2-ros-py==0.25.5
tf2-tools==0.25.5
theora-image-transport==2.5.0
toml==0.10.2
topic-monitor==0.20.3
tornado==6.1
trajectory-msgs==4.2.3
turtlesim==1.4.2
Twisted==22.1.0
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
visualization-msgs==4.2.3
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.0.2
xacro==1.13.3
zipp==1.0.0
zope.interface==5.4.0
